requirements.txt
selenium~=4.10.0
requests~=2.31.0
openpyxl~=3.1.2
twilio~=8.7.0
Pillow~=10.0.0
PyPDF4~=1.27.0
pdfplumber~=0.10.2
fpdf~=1.7.2
telegram~=0.0.1
telebot~=0.0.5